+++ CHANGELOG.rst
@@ -1,6 +1,21 @@
 Changelog:
 ================================================================================
 
+Release 1.2.0f0:
+================================================================================
+
+- PBSTLIOS-15: Updated script to work with Xcode 4.4 and Boost 1.50.0 based on https://svn.boost.org/trac/boost/ticket/6686.
+
+Sucessfully tested on the following configurations:
+
+- Mac OS X 10.8, Xcode 4.4 (Build 4F250), iOS SDK 5.1, Boost 1.50.0
+
+Notes: 
+
+- Support for ALL Boost versions prior 1_50_0 is dropped.
+
+- Boost is now compiled with Clang.
+
 Release 1.1.0f0:
 ================================================================================